Introduce a script to find staging deployable merge commits.
The script `qa-able-commits.sh` uses git and the hub command to list auto merge commits
between the current state of production and the current state of master.

This can be used to identify a suitable commit reference to deploy to staging for the
master QA cycle if there are too many changes to verify in a reasonable time frame.

Usage:
./qa-able-commits [NoRedInk repository path]
The path will be read from the environment variable LOCAL_REPO_PATH by default.